+++ src/extensions/surveys.tsx
@@ -365,7 +365,7 @@ export function Questions({
                                                 question,
                                                 idx,
                                                 survey.appearance || defaultSurveyAppearance,
-                                                (res) => onNextClick(res, idx),
+                                                (res) => onNextClick(res, question.questionIndex || idx),
                                                 () => closeSurveyPopup(survey, posthog, readOnly)
                                             )}
                                         </div>