+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+## 1.135.2 - 2024-05-24
+
+- feat: Add tests and fix regex for some obscure android browsers (#1196)
+
+## 1.135.1 - 2024-05-24
+
+- fix: Add missing retry info and test (#1202)
+- fix: Pass `retriesPerformedSoFar` to keep count of retries (#1201)
+
+## 1.135.0 - 2024-05-23
+
+
+
+## 1.134.0 - 2024-05-23
+- fix(survey): Retain question index if questions are shuffled (#1198)
+
+## 1.133.0 - 2024-05-22
+
+
+## 1.132.3 - 2024-05-22
+
+- feat(survey): support internal_targeting_flag to show/hide survey popup (#1181)
+
+## 1.132.2 - 2024-05-22
+
+- feat(web-analytics): Add function for explictly enabling person processing (#1191)
+
+## 1.132.1 - 2024-05-22
+
+- fix: redact not hide headers (#1190)
+
 ## 1.132.0 - 2024-05-13
 
 - chore: Bump release to 1.131.5 (#1183)